refactor(db): Split anime relation schema modules
